+++ src/emhass/command_line.py
@@ -11,10 +11,10 @@ import emhass
 from emhass.retrieve_hass import retrieve_hass
 from emhass.forecast import forecast
 from emhass.optimization import optimization
-from emhass.utils import get_yaml_parse, get_days_list, get_logger
+from emhass.utils import get_yaml_parse, get_days_list, get_logger, set_df_index_freq
 
 def setUp(config_path: pathlib.Path, base_path: str, costfun: str, 
-    params: str, logger: logging.Logger) -> dict:
+    params: str, set_type: str, logger: logging.Logger) -> dict:
     """
     Set up some of the data needed for the different actions.
     
@@ -24,6 +24,8 @@ def setUp(config_path: pathlib.Path, base_path: str, costfun: str,
     :type costfun: str
     :param params: Configuration parameters passed from data/options.json
     :type params: str
+    :param set_type: Set the type of setup based on following type of optimization
+    :type set_type: str
     :param logger: The passed logger object
     :type logger: logging object
     :return: A dictionnary with multiple data used by the action functions
@@ -33,40 +35,54 @@ def setUp(config_path: pathlib.Path, base_path: str, costfun: str,
     logger.info("Setting up needed data")
     # Parsing yaml
     retrieve_hass_conf, optim_conf, plant_conf = get_yaml_parse(config_path, params=params)
-    # Retrieve data from hass
-    days_list = get_days_list(retrieve_hass_conf['days_to_retrieve'])
-    var_list = [retrieve_hass_conf['var_load'], retrieve_hass_conf['var_PV']]
+    # Initialize objects
     rh = retrieve_hass(retrieve_hass_conf['hass_url'], retrieve_hass_conf['long_lived_token'], 
                        retrieve_hass_conf['freq'], retrieve_hass_conf['time_zone'], 
                        params, base_path, logger)
-    rh.get_data(days_list, var_list,
-                minimal_response=False, significant_changes_only=False)
-    rh.prepare_data(retrieve_hass_conf['var_load'], load_negative = retrieve_hass_conf['load_negative'],
-                    set_zero_min = retrieve_hass_conf['set_zero_min'], 
-                    var_replace_zero = retrieve_hass_conf['var_replace_zero'], 
-                    var_interp = retrieve_hass_conf['var_interp'])
-    df_input_data = rh.df_final.copy()
-    # Initialize objects
     fcst = forecast(retrieve_hass_conf, optim_conf, plant_conf,
                     params, base_path, logger)
-    df_weather = fcst.get_weather_forecast(method=optim_conf['weather_forecast_method'])
-    P_PV_forecast = fcst.get_power_from_weather(df_weather)
-    P_load_forecast = fcst.get_load_forecast(method=optim_conf['load_forecast_method'])
-    df_input_data_dayahead = pd.concat([P_PV_forecast, P_load_forecast], axis=1)
-    df_input_data_dayahead.index.freq=df_input_data.index.freq
-    df_input_data_dayahead.columns = ['P_PV_forecast', 'P_load_forecast']
     opt = optimization(retrieve_hass_conf, optim_conf, plant_conf, 
                        fcst.var_load_cost, fcst.var_prod_price, days_list, 
                        costfun, base_path, logger)
+    # Perform setup based on type of action
+    if set_type == "perfect-optim":
+        # Retrieve data from hass
+        days_list = get_days_list(retrieve_hass_conf['days_to_retrieve'])
+        var_list = [retrieve_hass_conf['var_load'], retrieve_hass_conf['var_PV']]
+        rh.get_data(days_list, var_list,
+                    minimal_response=False, significant_changes_only=False)
+        rh.prepare_data(retrieve_hass_conf['var_load'], load_negative = retrieve_hass_conf['load_negative'],
+                        set_zero_min = retrieve_hass_conf['set_zero_min'], 
+                        var_replace_zero = retrieve_hass_conf['var_replace_zero'], 
+                        var_interp = retrieve_hass_conf['var_interp'])
+        df_input_data = rh.df_final.copy()
+        # What we don't need for this type of action
+        P_PV_forecast, P_load_forecast, df_input_data_dayahead = None
+    elif set_type == "dayahead-optim":
+        # Get PV and load forecasts
+        df_weather = fcst.get_weather_forecast(method=optim_conf['weather_forecast_method'])
+        P_PV_forecast = fcst.get_power_from_weather(df_weather)
+        P_load_forecast = fcst.get_load_forecast(method=optim_conf['load_forecast_method'])
+        df_input_data_dayahead = pd.concat([P_PV_forecast, P_load_forecast], axis=1)
+        df_input_data_dayahead = set_df_index_freq(df_input_data_dayahead)
+        df_input_data_dayahead.columns = ['P_PV_forecast', 'P_load_forecast']
+        # What we don't need for this type of action
+        df_input_data = None
+    elif set_type == "publish-data":
+        df_input_data, df_input_data_dayahead = None
+        P_PV_forecast, P_load_forecast = None
+    else:
+        logger.error("The passed action argument and hence the set_type parameter for setup is not valid")
+
     # The input data dictionnary to return
     input_data_dict = {
         'root': base_path,
         'retrieve_hass_conf': retrieve_hass_conf,
+        'rh': rh,
+        'opt': opt,
+        'fcst': fcst,
         'df_input_data': df_input_data,
         'df_input_data_dayahead': df_input_data_dayahead,
-        'opt': opt,
-        'rh': rh,
-        'fcst': fcst,
         'P_PV_forecast': P_PV_forecast,
         'P_load_forecast': P_load_forecast,
         'costfun': costfun,
@@ -204,6 +220,7 @@ def main():
     parser.add_argument('--costfun', type=str, default='profit', help='Define the type of cost function, options are: profit, cost, self-consumption')
     parser.add_argument('--log2file', type=bool, default=False, help='Define if we should log to a file or not')
     parser.add_argument('--params', type=str, default=None, help='Configuration parameters passed from data/options.json')
+    parser.add_argument('--runtimeparams', type=str, default=None, help='A dictionnary of runtime optimization parameters')
     parser.add_argument('--version', action='version', version='%(prog)s '+version('emhass'))
     args = parser.parse_args()
     # The path to the configuration files
@@ -212,9 +229,9 @@ def main():
     # create logger
     logger, ch = get_logger(__name__, base_path, save_to_file=args.log2file)
     # Setup parameters
-    params = args.params
-    input_data_dict = setUp(config_path, base_path, args.costfun, params, logger)
+    input_data_dict = setUp(config_path, base_path, args.costfun, args.params, args.action, logger)
     # Perform selected action
+    runtimeparams = args.runtimeparams
     if args.action == 'perfect-optim':
         opt_res = perfect_forecast_optim(input_data_dict, logger)
     elif args.action == 'dayahead-optim':