+++ tests/test_optimization.py
@@ -82,6 +82,47 @@ class TestOptimization(unittest.TestCase):
         now_precise = datetime.now(self.input_data_dict['retrieve_hass_conf']['time_zone']).replace(second=0, microsecond=0)
         idx_closest = self.opt_res_dayahead.index.get_indexer([now_precise], method='ffill')[0]
         idx_closest = self.opt_res_dayahead.index.get_indexer([now_precise], method='nearest')[0]
+        # Test the battery
+        self.optim_conf.update({'set_use_battery': True})
+        self.opt = optimization(self.retrieve_hass_conf, self.optim_conf, self.plant_conf, 
+                                self.fcst.var_load_cost, self.fcst.var_prod_price,  
+                                self.days_list, self.costfun, root, logger)
+        self.opt_res_dayahead = self.opt.perform_dayahead_forecast_optim(
+            self.df_input_data_dayahead, self.P_PV_forecast, self.P_load_forecast)
+        self.assertIsInstance(self.opt_res_dayahead, type(pd.DataFrame()))
+        self.assertTrue('P_batt' in self.opt_res_dayahead.columns)
+        self.assertTrue('SOC_opt' in self.opt_res_dayahead.columns)
+        self.assertAlmostEqual(self.opt_res_dayahead.loc[self.opt_res_dayahead.index[-1],'SOC_opt'], self.plant_conf['SOCtarget'])
+        
+    def test_perform_naive_mpc_optim(self):
+        self.df_input_data_dayahead = self.fcst.get_load_cost_forecast(self.df_input_data_dayahead)
+        self.df_input_data_dayahead = self.fcst.get_prod_price_forecast(self.df_input_data_dayahead)
+        # Test the battery
+        self.optim_conf.update({'set_use_battery': True})
+        self.opt = optimization(self.retrieve_hass_conf, self.optim_conf, self.plant_conf, 
+                                self.fcst.var_load_cost, self.fcst.var_prod_price,  
+                                self.days_list, self.costfun, root, logger)
+        prediction_horizon = 10*self.retrieve_hass_conf['freq']
+        soc_init = 0.4
+        soc_final = 0.6
+        past_def_load_energies = [3000, 750]
+        self.opt_res_dayahead = self.opt.perform_naive_mpc_optim(
+            self.df_input_data_dayahead, self.P_PV_forecast, self.P_load_forecast, prediction_horizon,
+            soc_init=soc_init, soc_final=soc_final, past_def_load_energies=past_def_load_energies)
+        self.assertIsInstance(self.opt_res_dayahead, type(pd.DataFrame()))
+        self.assertTrue('P_batt' in self.opt_res_dayahead.columns)
+        self.assertTrue('SOC_opt' in self.opt_res_dayahead.columns)
+        self.assertAlmostEqual(self.opt_res_dayahead.loc[self.opt_res_dayahead.index[-1],'SOC_opt'], soc_final)
+        self.assertAlmostEqual(self.opt_res_dayahead['P_deferrable0'].sum()*(
+            self.retrieve_hass_conf['freq'].seconds/3600), 
+                               self.optim_conf['P_deferrable_nom'][0]*self.optim_conf['def_total_hours'][0]-past_def_load_energies[0])
+        soc_init = 0.8
+        soc_final = 0.5
+        self.opt_res_dayahead = self.opt.perform_naive_mpc_optim(
+            self.df_input_data_dayahead, self.P_PV_forecast, self.P_load_forecast, prediction_horizon,
+            soc_init=soc_init, soc_final=soc_final, past_def_load_energies=past_def_load_energies)
+        self.assertAlmostEqual(self.opt_res_dayahead.loc[self.opt_res_dayahead.index[-1],'SOC_opt'], soc_final)
+        
         
 if __name__ == '__main__':
     unittest.main()